GNUmakefile:  Clone macward-compat for ams-quartz
Even though we're not using QuickDraw (or any other 64-bit-anathematized
APIs), Nitrogen (a dependency of macvx) expects either QuickDraw headers
to be available as in 32-bit builds in 10.6, or missing-macos headers as
a substitute.  It would be nice to rid Nitrogen of this requirement, but
in the meantime, solve the issue by explicitly declaring the dependency.