Revert "ams-ui:  Clip TEScroll() rect to stay in viewRect"
This was the wrong fix and caused a regression in Deja Vu's text window,
making ScrollRect() effectively a no-op (by clipping the srcRect down to
an empty rect) and resulting in scrolled-in text drawing over prior text
that was meant to have been scrolled out but didn't actually get erased.

This revives (temporarily) the text scrolling bug in Risk's help dialog.

This reverts commit 682845113b3b7a01bbfcee251e26c2d81a0a5c01.