ams-qd:  Rename ScrollRect()'s srcRect parameter
This is a misnomer that led to confusion over what ScrollRect() actually
does.  The source bits are everything visible in the provided rectangle,
and the destination bits are whichever bits /within that same rectangle/
that the scrolled source bits map to -- everything outside gets clipped.

(If you want to copy a set of bits from here to there, call CopyBits().)