LRMusicFix:  Modify wave table, not phases
Instead of patching VInstall() to assign a phase of 64 (which is correct
for the triangle waveform that Lode Runner uses), rotate the waveform to
align it with a phase of 0.  This strategy is simpler if we're intending
to support Mac OS as well as Advanced Mac Substitute; in Systems without
the Process Manager, we'd additionally have to patch ExitToShell() so we
could ensure removal of the VInstall() patch on exiting the application.