amicus:  Disassociate mouse from cursor position
This is the proper way to do it, and fixes the intermittent issue of the
native cursor sometimes failing to stay hidden in macOS 10.13 and later.

Since our current approach relies on CGEvents, but CopyEventCGEvent() is
only available since 10.5, only use the disassociated mode in 10.5+, and
keep using an associated cursor and CGWarpMouseCursorPosition() in 10.4.