Compile::Driver:  Quell 10.13 linker warning spam
The system-provided SDK (in /System/Library/Frameworks) in "High Sierra"
(macOS 10.13) is broken with regard to linking, in two ways.  First, the
framework link targets are provided in two different forms (conventional
shared library stubs and .tbd files, i.e. a text-based definition) which
are out of sync with each other.  (Curiously, it's the files in the more
recently added format that are out of date.)  This generates unavoidable
linker warnings, in massive quantities, when linking against frameworks.

The second issue applies only to 32-bit builds, but to all link targets,
including those not linking against any framework nor dependent on macOS
in any way.  We're told "The i386 architecture is deprecated for macOS",
for every 32-bit executable that gets linked, again with no mechanism by
which to disable it.  The officially sanctioned solution of switching to
another SDK by setting SDKROOT is unusable in the default case -- if the
command-line tools are installed automatically in macOS 10.13, they come
with only the 10.14 SDK, which lacks 32-bit support.  That doesn't work.

Apple promised that 10.13 would run 32-bit code without compromises, but
they've sabotaged the means of building it in the first place.  While it
remains technically possible to successfully produce the executables, it
comes at the cost of pages of linker warnings.  Perhaps Apple views this
as punishment for using the computer you gave them money for in a manner
they find unorthodox.  I see it as callous, mediocre, and short-sighted.

In any case, it's an unacceptable inconvenience for those developing the
various applications that run just fine in 10.13, and for users building
them from source.  In response, I'm taking the drastic step of silencing
all linker warnings when linker warning spam is predicted.  This affects
only builds using the 10.13 SDK, with SDKROOT unset, when frameworks are
linked or when building for the 32-bit i386 architecture.  I wish such a
measure weren't warranted.  However, my other efforts at mitigation have
not been successful.  My first plan to address the framework-out-of-sync
warnings ultimately worked, but only after adding substantial complexity
to Compile::Driver -- and without affecting the architecture deprecation
warnings.  My second effort, to capture and filter linker output, failed
entirely (though such a scheme should find more success in A-line, which
already does so for classic Mac OS builds).  I have no other ideas left.

One workaround is to install Xcode 9.x, which includes the 10.13 SDK; in
that case, SDKROOT will be set and linker warnings will not be silenced.