ams-fs:  Ignore volume names in absolute paths
This is one of the two changes needed so that The Fool's Errand v1.1 can
find its Fool's Puzzles data file.  It seems like it would be preferable
to consider only VCBs whose volume name matches the one in the pathname,
but doing so actually breaks The Fool's Errand, which might be depending
on the Poor Man's Search Path or some other quirk that's yet unemulated.