ams-fs:  Match "Volume:" in by-name VCB_lookup()
This is the second of two changes required for The Fool's Errand v1.1 to
find its Fool's Puzzles data file.  Without this change in place, you're
asked to locate the file yourself; though with only this change applied,
the game fails to find the file and won't even let you make the attempt.