ams-qd:  Fix picture pattern origin offset bug
Since we set the origin to draw in picture frame coordinates, compensate
by adjusting patterns to use the same origin as with window coordinates.

This fixes a minor graphics glitch drawing the background pattern in The
Fool's Errand when a tile is moved in The Star, The Chariot, and others.