rbuild:  Ensure NEXUS_HOST is set ahead of time
This will avoid problems if the hostname changes during a build.  (Which
is not merely theoretical -- a switch between wifi and ethernet when the
DHCP servers (or one DHCP server and the host) disagree on the name will
result in a sudden switch from one nexus subdirectory to another, almost
certainly while nexus-call is waiting for its result.  In that case, the
event is harmless to the build and merely logs a warning and leaves file
detritus lying around, but it is theoretically possible for it to switch
host names between nexus-new and nexus-call, causing the latter to fail.