Genie:  Make IterateIntoCache_Completion() static
Since we're no longer using IterateIntoCache_Completion() as a parameter
in a template, it no longer requires extern linkage.  Declare it static,
move it out of the anonymous namespace, and axe the __MACH__ conditional
guard that was only there because 10.13's clang somehow thought, against
all logic and reason, that the function actually had **static** linkage.

This reverts commit fb6dae56b735f24b4abd7c6f0b6bc9c952c931ad.