[DEL] Axe hapless 68K alloca() implementation
It turns out that the name `__alloca()` was load-bearing, and MWC68K was
providing an implementation all along.  It's only the /declaration/ that
MWCPPC provides internally and MWC68K doesn't.  The C compiler has to be
in on the scheme so it can save and restore the SP in the prolog/epilog.

This fixes the regression from a3dc46a8cde9d11b4ac86696c2ec6df6a214e6fd.