MacFixup, mwdebug:  Include MixedMode.h explicitly
This frees UPP.h of the task of including <MixedMode.h>, and with it the
burden of choosing a form of it that must satisfy all clients.  At issue
is the need of some clients (e.g. Nitrogen) to use UPP types as template
parameters, which requires that they be distinct types.  The new project
named "unique-CFM-UPPs" contains a MixedMode.h header that will override
the one in Universal Interfaces, making UPP types distinct even for CFM.