MacFixup:  Fix invoker in DEFINE_UPP_2() macro
This error went unnoticed because the problematic code resulting from it
(a) passed the wrong number of arguments to a /variadic/ function (which
rules out prototype matching), and (b) never got called -- the UPP types
in question (ThreadSwitch and ThreadTermination) are only created, never
invoked.  (Actually, ThreadTermination doesn't even get created at all.)
Incidentally, those two cases are still wrong, because they return void.

We basically don't use CFM-68K, and there's almost no reason to call one
of the InvokeFooUPP() functions here -- we're providing the callbacks to
the OS, not calling them -- but assuming we do both, it'll be a problem.