v68k-screen, xv68k:  Copy to transit all at once
Call memcpy() to populate the transit buffer once per update rather than
once for each write to screen memory.  This offers two primary benefits:
One, the memcpy() call for /every/ write to screen memory (even when the
screen is locked for updates) is eliminated.  The fullscreen memcpy() to
the transit framebuffer isn't an added expense -- it merely replaces the
memcpy() to the previous_frame buffer which had been used for comparison
to determine if an update is needed, for which the virtual buffer is now
used instead.  Two, the transit framebuffer now remains unmodified until
an update occurs, which could reduce tearing when updates occur rapidly.