v68k-screen:  Rename buffer var to transit_buffer
The term "screen buffer" is ambiguous.  We have both the mmapped backing
store which is shared with the front end (henceforth a "transit buffer")
and optionally a separate, unshared screen buffer that's just some chunk
of a larger segment at a known address in guest memory (now to be called
a "virtual buffer").  They correspond roughly to VRAM and the usual RAM,
respectively.  If the xv68k --mem option is passed, then virtual buffers
exist in guest memory at fixed addresses, and the transit buffers aren't
directly accessible.  (Otherwise, the virtual buffers don't exist, while
guest memory maps to the transit buffers, but this usage is deprecated.)