ams-core:  Fix {Hand,Ptr}AndHand() of 32K+ handles
Because the trap dispatcher invariably does `TST.W D0` before returning,
we have to explicitly `TST.L D0` after invoking _GetHandleSize, so as to
correctly test the entire longword value.  Otherwise, if a handle's size
mod 64K is at least 32K, the error test will fail with a false positive.

HandToHand() was already correctly testing the full longword result, but
HandAndHand() and PtrAndHand() were flawed.  If you wanted an absolutely
stripped-to-the-frame build, you could drop the conditional branches (as
well as the preceding longword tests) and still have code that works for
valid handles only.  But we see what some applications do in the dark...