skif:  Axe 'cast' subcommand
The distinction between cast and bump has always been unclear.  It would
appear that 'bump' is meant ot increment the seed while 'cast' is not --
but they both call raster::broadcast(), which bumps the seed regardless.
In any case, the one other job of raster::broadcast() was to broadcast a
process-shared condition variable, but we did away with that in January,
leaving raster::broadcast() as a glorified increment operator.  It makes
no sense to have two commands doing essentially the same thing, thus one
of them now gets removed, and 'cast' (which is short for "broadcast") no
longer has a relevant meaning (since the condition variable is gone), so
we'll drop that one and keep 'bump', which will now increment just once.