Genie:  Empty-check resource type count, not EOF
This makes us dependent on having at least a 128K ROM (or the equivalent
system software), but we don't run in 512K anyway, so it's an acceptable
limitation.  Whereas GetEOF() works as far back as System 1.0, but since
System 6 doesn't automatically compact a resource file when removing the
last resource like System 7 does, we need an alternative emptiness test.