glfb-common:  Always use rectangle textures
We already use GL_TEXTURE_RECTANGLE_ARB when available, which allows for
arbitrary texture rectangle dimensions, like 640x480.  However, it's not
always available (as in Mac OS X 10.2).  In that case we've been falling
back to GL_TEXTURE_2D, which is always present, but has some differences
such as using a normalized texture coordinate space (with coordinates in
the range [0.0, 1.0]) instead of counting pixels.  Fortunately, there is
a middle ground:  The GL_TEXTURE_RECTANGLE_EXT texture type shares pixel
coordinate space with GL_TEXTURE_RECTANGLE_ARB and is available in 10.2,
with the only drawback being that it doesn't allow arbitrary dimensions.