vlib:  Explicitly reserve `a+b` and `a-b` syntax
Here, `a-b` should be read as an expression whose `a` isn't literally an
identifier, as hyphens are valid in identifier names.  But `3-4` used to
evaluate to -1, for example, and now it's an error.  However, the reason
it's an error is that it gets parsed as `3(-4)`, and an implied operator
for integers doesn't exist yet.  If it did, `3-4` would evaluate to -12!

This new change preempts that parsing and makes it an explicitly  error.