xv68k:  Darken screen (both pages) even sans --mem
We were prepainting the screen to black only when a non-traditional size
was used or when --mem was specified to set MemTop and set screen buffer
addresses below RAM, but not in the default case of a 512x342 monochrome
screen in NuBus space.  While we're here, also darken the page 2 buffer.