[NEW] v68k-record:  Instrumenting library for v68k
Initially conceived to enable the creation of screencasts (involving the
capture of video and audio signals), the scope of this project increased
when using a (simulated) hardware cursor became an option (as the cursor
shape, location, and other details aren't reflected in the screen image,
and would have to captured separately).  Similarly, it could be valuable
to record user input (e.g. which keys are being pressed) to help viewers
understand what's happening on screen when shortcut keystrokes are used.

Finally, this facility could further be used for optionally logging low-
level information like VBL interrupts or accesses to Ticks, to assist in
debugging timing problems like audio gaps or premature graphics updates.