MacBinaryLib:  Fix encode of nested directories
Our preemptive clearing of ioNamePtr in order to prevent errors that are
challenging to reproduce and fiendishly difficult to debug led, in part,
to an error that's 100% reproducible and trivial to debug -- not too bad
a tradeoff, methinks.  Copy the list entry to the stack so it can't move
in memory, and repopulate the ioNamePtr field (pointing into the stack).

This fixes the regression from f9a196114bdb738a516027a6899d08507b200113.