mac-app-utils:  Plot 'ICN#' with white background
Drawing an About box with a black-and-white icon got the grey background
even within that area of the icon included in the mask.  Paint it white.

This is a different issue than 17d7661637803d610939991d24756590e2709306,
with a different solution.  In our case, calling `BackColor( qd.white )`
was ineffective, and calling RGBBackColor() was necessary.  (By shifting
the plot_icon_id() call to the end, we don't need to restore the color.)