ams-io:  Use VBL, not timer, to complete four-tone
It doesn't make sense to schedule a timer for four-tone sound completion
when it's vendor-recommended practice to dynamically alter the duration.
Since we're using a VBL task to monitor the FTSoundRec for such changes,
we can check for the duration reaching zero when the VBL task proc runs.
By keeping SoundVBL.vblCount non-zero until the I/O request's completion
routine exits, we can use it as an indication of whether the VBL task is
installed or not.  (Alternatively, we could defer the completion routine
call until we've left the VBL task loop, via the Deferred Task Manager.)

This prevents the removal and reinstallation of SoundVBL when an initial
duration elapses after having been dynamically extended, the potentially
delayed removal of SoundVBL after the sound ends, and a complete failure
to call the I/O request's completion routine after the duration expires.