postlink-68k-standalone:  Use BlockMove vs. HWPriv
This makes the runtime 18 bytes smaller, and it only makes one trap call
rather than two (to check for _HWPriv) or three (if it's present).  Also
of note is that two internal branches are eliminated.  Furthermore, it's
conceivable that an implementation of BlockMove might not invalidate all
instruction cache lines, but only those intersecting the target range of
the move operation (as opposed to the entire code cache as HWPriv does).