amicus:  Initialize 'sync' note contents to zero
If the seed is initialized to -1, and the local copy is zero, and if the
front end bumps the shared seed to zero before we get a chance to check,
then the seeds will match and we'll miss the first update.  In any case,
zeroing the relay object is consistent with the behavior of `skif make`.