amicus:  Initialize cursor image to zero, not ones
This is important for two reasons:  One, should the cursor get displayed
before the image is set, it will be completely transparent rather than a
solid black square.  Two, the cursor seed in the third frame is properly
initialized to zero, not -1 (which when bumped to zero could mislead the
front end into ignoring the first update).  The latter issue has in fact
been observed in Amphitheatre (though not in any other macOS front ends,
for reasons that remain unclear).