Tic-tac-toe:  Select game window going fullscreen
Before going fullscreen, select the game window.  This is important when
a desk accessory is frontmost, as that allows menu item selection (which
lets us enter fullscreen mode) but intercepts keystrokes (so hitting Esc
or Cmd-F (or Cmd-W, or Cmd-Q...) doesn't let us /exit/ fullscreen mode).

After selecting the window, clear the update region to prevent a useless
(and counter-productive) update event.