ams-qd:  Pseudo-close open polygons in PolyRgn()
For an unclosed polygon, re-append the first point without adjusting the
stored polySize, letting us to iterate through pairs of adjacent points.

This logic already existed in StdArc() (by way of clip_from_angle()) and
ClosePoly() (and therefore, indirectly, CloseRgn()).  However, StdPoly()
lacks it, so in some cases we'd be reading off the end of the PolyHandle
data.  Consolidate the logic into PolyRgn(), where it's actually needed.

In practice, the polygons passed to StdPoly() are constructed by calling
OpenPoly/.../ClosePoly (preventing an issue), but this isn't guaranteed.