ams-ui:  Don't release dialog, item list resources
Inside Macintosh, Volume I documents that GetNewDialog() doesn't release
resources and that it makes a copy of the item list resource; therefore,
it's incorrect for us to call DetachResource() as that's effectively the
same as calling ReleaseResource() after making a copy.  The routines for
alerts are less clearly described, but presumably they behave similarly.