QDFullscreen:  Use a window vs. a GrafPort sans QT
If QuickTime 2.5 isn't available for BeginFullScreen(), we manually open
our own fullscreen graphcis port.  But instead of using a bare GrafPort,
make it a proper window, so that (a) it covers other windows (preventing
title changes from clobbering the screen) and (b) it gets update events.
Both of these are important for document management -- we need the first
so we can update a document window's title on the fly, and the second to
redraw regions of the screen after dismissing the Standard File dialogs.