vfs:  Deimplement posix_slurp()
This was only needed by freemount-server in 2013, until posix_open() was
added.  At that point, freemount-server's read handler would call open()
first, and only if open() failed with ENOENT would it then call slurp(),
which would necessarily fail in the same way.  (This logic was added for
supporting virtual files that can be slurped but not opened, which isn't
a phenomenon occuring with POSIX nodes -- only with Mac metadata files.)