MacBinaryLib:  Explicitly use plus::var_string
This is a physical no-op for Relix builds.  For Mach-O, it switches from
std::string to plus::var_string for smaller code (behaviorally a no-op).

Also, bring FSIORefNum into scope for later.  The next change only needs
it in MacBinary.cc, but a subsequent patch will expect it in the header.