Nitrogen:  Axe ptmv-template-param-dependent code
These are template variations of FSGetCatalogInfo(), FSSetCatalogInfo(),
FSGetVolumeInfo(), and FSSetVolumeInfo() whose implementation depends on
using pointers to member variables as template paremeters.  But we don't
have that feature in Metrowerks C++ 2.4, the last version to support 68K
development.  While it's a theoretical option to use them in Mach-O-only
code, or even to use a later Metrowerks compiler for PPC, it's not worth
the complexity cost.  (Consider, for example, that each variant can only
get or set one piece of information per call to the underlying routine.)