Vertice:  Extend window to bottom in Carbon builds
We were starting with the small window size to avoid the failure mode of
running out of memory before displaying the window, denying the user the
ability to mitigate the issue by resizing the window.  In Mac OS X, this
is much less of a concern, and also screens tend to be rather larger, so
a larger window size is appropriate.