Genie:  Rename/detemplatize FSTreeFromFSDirSpec()
We no longer deal with the FSDirSpec type, just VRefNum_DirID.  Somehow,
making it a non-template generates smaller code in Metrowerks C++ (aside
from the smaller symbol name).  It does this even without the new header
include that clang now requires to compile at all.  (Clang wants to know
the details of intrusive_ptr so it can implement the copy that it's just
going to elide anyway, which I'm essentially forced to code in the first
place because C++ has no hygienic macros, not because I want it copied.)