MacScribe:  Branch less in make_quad_name()
The common case is d != ' ', which will now execute straight through the
switch instead of potentially branching at three different points, which
in theory could improve performance on some architectures.  In practice,
this code is only used for human display and isn't performance-sensitive
in the slightest, but this change does make the code a tiny bit smaller.