[NEW] Pearl:  SDL-based, windowed front end to AMS
Pearl behaves very similarly to -- and is in fact based upon --
the existing amicus-based front ends.  Its main advantage over
its inspiration is that it uses the cross-platform SDL2 library
<https://wiki.libsdl.org/SDL2/>, allowing it to target non-Mac
platforms.  It also includes some extra niceties such as run-time
toggling of integer bitmap scaling.

Pearl integrates libsndtrack with SDL2's audio facilities, obviating
the need for a separate sndtrack process (and a PortAudio dependency).

Pearl is designed as a drop-in replacement for the amicus-based front
ends + sndtrack bundled using Retromatic.  On non-Mac platforms, Pearl
expects a directory hierarchy and contents similar to that of a
Retromatic-produced macOS .app bundle.

A MainMenu.xib file is provided defining a menu similar to that used by
Carbon-based front ends using aglow.  For macOS targets >= 10.8, this
optionally can be built and included in Pearl's app bundle to replace
the default SDL-provided menu.