Chess:  Work around OS X 10.4+ Pattern origin bug
When calling FillRect() to draw squares on the board, the pattern always
begins at the eighth row (followed by the first row, then second, etc.),
rather than respecting the port's origin.  This behavior was observed to
be present in Mac OS X 10.4 and later (specifically, 10.4 on PPC as well
as 10.6 through 10.9, 10.12, and 10.13 on x86), while absent in Mac OS X
10.2 and 10.3, regardless of compiled runtime/architecture -- the bug is
present when running a Carbon/PPC/CFM build in Rosetta in Mac OS X 10.6.