Compile::Driver:  Don't set Mac OS X version macro
By default, MAC_OS_X_VERSION_MIN_REQUIRED will match the SDK being used.
We only need to override this in order to produce a build that would run
on an older system, which (a) isn't strictly necessary when we can build
from source, and (b) would require run-time checks for code symbols that
are currently only guarded with compile-time checks.