Nitrogen, Genie:  Swap type/creator in FSSignature
This is the canonical ordering: type, then creator.  It's in FInfo, it's
in PkgInfo, and it's on the stack when you call a routine like HCreate()
(to which you pass the creator first positionally, but since it's pushed
first (due to `pascal` calling convention), it ends up higher in memory,
so the pattern holds) -- at least on 68K, anyway.