amicus:  Send kEventAmicusNewPalette events
Notify the render thread that the CLUT has been modified, so the palette
must be updated.  We need this now even without palette animation due to
a race condition between spawning the xv68k back end (which will set the
palette when the Graphics Devices module is initialized), and making the
initial call to set_palette() in the front end.  (We'll also need it for
palette animation).