ams-qd:  Axe redundant Rect intersection logic
The portBits.bounds intersection is redundant because it also happens in
get_refined_clip_region().  The intersection with portRect doesn't occur
there (see 9d384d909e65bb52ab95636d88cc9c3fcecd37c1), so we can't remove
it, but we can perform the operation in advance (which may theoretically
reduce the cost of the clipRgn intersection in some cases).  Also, using
SectRect() instead of a hand-coded substitute obviates the need for one.