display-fbdev:  Re-update now if new raster seed
If a graphics update occurred while we were busy processing the previous
one, handle it immediately rather than erroneously waiting for the next.

This fixes an issue in Advanced Mac Substitute on a Raspberry Pi 3 where
closing a menu issued one update to unhighlight the menu title, and then
a second update to hide the menu.  Since the blit that restores the bits
under the menu is mostly native code, it's very quick (a bunch of memcpy
operations, basically), while the front end transcodes an entire screen.
The second notification got lost while we were processing the first, and
proceeding to wait for a third update (before this patch) left the stale
menu visible until such an update occurred (typically, cursor movement).