Genie:  Use hide/show over inval/draw for list fix
Prefer HideControl() and ShowControl() (which are universally available)
to InvalRect() (which requires using a compatibility function in Carbon)
and DrawOneControl() (which was added in the 128K ROM).  While depending
on the 128K ROM is irrelevant here (since the List Manager does anyway),
it's still valuable to promote the principle of minimizing dependencies.

This change will have a slight cosmetic effect when resizing a list with
one scroll bar that would overlap the window's grow box:  The scroll bar
(including the grow box area) is erased immediately, whereas previously,
the downward scroll arrow would remain visible after redrawing the newly
adjusted scroll bar (until DrawGrowIcon() is called at the next update).