ams-core:  Prefer get_Ticks() call to Ticks access
In preparation for future measurement and potential throttling of access
to Ticks by application code, avoid confounding them with our own use of
Ticks in the OS Event Manager and in Delay().  Call get_Ticks() instead,
which draws from the same source as Ticks but technically doesn't access
it.  Leave TickCount() as it is -- it's effectively just a read of Ticks
by way of the trap dispatcher (and slower).  We don't call it ourselves.