ams-common:  'SICN'-fix large_icon_key() predicate
Menu items with a 'SICN' resource for the icon were being misreported as
having a large icon.  A new MDEF-0 build is required to draw those items
with the correct height, and a new ams-ui is needed to properly refigure
the menu size when swapping between a large icon and a 'SICN' (if ever).