vlib:  Downgrade unary op internal error to throw
This path is reached when trying to `begin` a struct (and potentially by
other means yet undiscovered).  Until we have a better response for this
issue, let it suffice to throw an exception that at least reports a line
number and can be caught by the executing program via a try/catch block.