tlsrvr:  Don't send an empty command to ToolServer
Previously, the command was written to a file, whose filename became the
actual command sent to ToolServer.  Now, the original command is sent in
the Do Script event.  Either way, the empty command precedes syntax that
redirects input, but in the latter case, the I/O redirection is mistaken
for a command and an error is reported, whereas in the former case, it's
not.  Why this is remains a mystery, but in any case, nix the uncommand.

This fixes the regression from 6c06e701575dd4f655acc419c803e44503211f02.