A-line:  Circumvent the include union if unusable
To reduce the number of include directories that have to be searched for
many different files being compiled, we maintain a single directory that
contains symlinks to projects' subdirectories (if they've been indicated
as being unique in the A-line config file, via the `sources` directive).

However, if the C preprocessor can't handle symlinks, we have to refrain
from using the shared include directory.  This is the case when we're in
Mac OS X on PPC, targeting the Metrowerks compiler running in ToolServer
(within Classic), and we're letting the compiler do the C preprocessing.

If we use a separate tool like mxcpp to preprocess in advance, it'll run
natively and handle the symlinks just fine, producing a translation unit
in a single file with no includes.  On the other hand, if we're building
in MacRelix, then symlinks are actually aliases under the hood[1] (which
the compiler can resolve on its own, without heroic effort on our part).

[1] Symlinks created in System 7 (or later) to HFS items that existed at
the time the symlink was created are made into aliases (*in addition* to
being symbolic links).  It's complicated (and beyond this text's scope).