ams-ui -> ams-seg:  Move Scrap Manager routines
For a Toolbox API, the Scrap Manager is fairly low-level.  While most of
ams-ui's components are layered on QuickDraw, and nearly all of them use
resources, the Scrap Manager needs only a handle and file I/O -- even if
much of what it stores is in the same format as resource data elsewhere.

Moreover, there's a dependency issue:  ExitToShell() will be modified in
the near future to call ShutDownPower(), which in turn (as documented in
Inside Macintosh, Volume V) calls UnloadScrap().