ams-seg:  Load scrap temporarily during get/put
PutScrap() and GetScrap() are supposed to operate on the scrap whereever
it happens to be stored, whether in memory or on disk.  This flexibility
catered to memory-constrained systems, like the original Macintosh (with
its 128K of RAM), allowing clipboard operations on data too large to fit
in memory when duplicated.  Four decades later, however, it's simply not
a major concern anymore.  That said, we still need to be compatible with
applications expecting to find the scrap in the Clipboard File, so while
we'll just load the scrap into memory when we need it, we'll also unload
it back to disk when we're done with it.