README.md:  Drop mention of the GPL (leaving AGPL)
My original intent was that some things would be released under the GPL,
and some other things under the AGPL, and which license applied to which
project would be specified somewhere -- though but for a few exceptions,
that hasn't been the case.  Even had it been so, the statement is easily
construed to mean that every component not otherwise labeled as released
under both the GPL and the AGPL (even though this would be redundant, as
the GPLv3 permits conversion to the AGPLv3).  To resolve this ambiguity,
and since I see no reason to forgo the additional protection afforded by
the AGPL in the general case, and since I always have the option to name
the GPL (or any other license) as a license for individual components as
need arises, I'm removing mention of the GPL here as a relevant license.