ams-rsrc:  Potentially extend file in SetResInfo()
If setting a new name that's longer than the previous, extension occurs.

Unless setting the name fails, set the mapChanged attribute, even if not
changing the name or the ID, even though Inside Macintosh, Volume I says
that a call to ChangedResource() is required to have the changes written
to the resource file.  Everything from System 6 to OS X works like this.