ams-ui:  Don't oscillate in TextEdit autoscroll()
When the TextEdit view rect is vertically shorter than one line of text,
successive calls to autoscroll() (e.g. upon each keystroke) were causing
the view to scroll alternately to the top and to the bottom of the line.

This was seen writing a name on Deja Vu's certificate, but is now fixed.