CDEF-1:  Fix drawing of inactive scroll bars
Sometimes we're called by HiliteControl(), in which case param will have
the value kControlInactivePart, so we must recognize that in addition to
kDrawControlIndicatorOnly and 0 (the entire control, rendered normally).
But in other cases (e.g. DrawControls()), we're called with param set to
0, so we have to check the contrlHilite field to know if we're inactive.

(See also 6b1a52366aab638574413149f300b5fd231f01c8.)