[NEW] Loader:  An Advanced Mac Substitute launcher
The Loader provides a means for launching either an application named by
the environment (e.g. via `AMS_OPEN=APPL:$app_name`) or the first one it
finds in a search of all volumes.  It displays the name and icon of that
application (which might not ever be shown while running the application
itself), providing important context for screencasts.  Additionally, the
Loader arranges to be relaunched when the application exits -- providing
"bookends" (a recognizable return to one's starting point).  Finally, in
the event that an application exits prematurely, the Loader gives a user
the option of relaunching it without restarting Advanced Mac Substitute.