ams-snd:  Add --max-channels option
This exists for two reasons:  One, so we can pass --max-channels=1, thus
preventing an application (like TaskMaker) from trying to play sounds on
multiple channels simultaneously (instead of interrupting a sound on the
one channel to start a new sound, for example).  Two, so we can use zero
as the maximum -- when applications unconditionally call SndNewChannel()
but fall back to using the Sound Driver, --max-channels=0 achieves that.
(If _SndNewChannel is trap-tested first, we can leave ams-snd unloaded.)