ams-snd:  Handle bufferCmd in SndDoImmediate()
It makes sense to call SndDoImmediate() for flushCmd and quietCmd (which
is a widely-used pattern).  These commands are administrative in nature,
and so must be executed immediately:  Culling the queue requires jumping
the queue, and prematurely ending the current sound mandates not waiting
for it.  These commands are also atomic -- they run instantly (or appear
to, at least).  In stark contrast is bufferCmd, which by its very nature
will take some measurable length of time.  SndDoCommand() is the perfect
tool for that job, and SndDoImmediate() far from it.  And yet, TaskMaker
chooses the latter.  Perhaps this is meant as an example of making one's
own choices rather than blindly following the instructions you're given?

In any case, TaskMaker passes a bufferCmd to SndDoImmediate() and Mac OS
copes with it.  How Advanced Mac Substitute should handle it isn't quite
clear, as the Sound Manager documentation is uncharacteristically silent
on the topic.  This implementation conservatively forwards the bufferCmd
to SndDoCommand() only if the queue is empty (ensuring that the buffered
sound will in fact start playing immediately), and rejects it otherwise.