ams-snd:  Bump the max payload length to 81 frames
We can play a sound that's too long to fit in a single message buffer if
we split it into segments.  But we only do so if the sample rate already
matches the 22 kHz output sample rate (so that no resampling occurs) and
we can be sure that the boundary between messages won't yield an audible
artifact.  Therefore, messages should contain a multiple of 370 samples,
ensuring no partial frames (except possibly the last one), not exceeding
30000.  We're close at 29600 (370 * 80), but 29970 (370 * 81) is closer.