MacBinaryLib:  Add option to clear all Inited bits
Clearing the kHasBeenInited bit loses the icon position, but maintaining
it prevents the Finder from adding BNDL icons to the desktop database --
so you're damned either way.  The default will continue to be the latter
(since a stale desktop database can be corrected, while the former loses
information permanently), which may be preferred for archiving one's own
files, but for new installations, updating the desktop database might be
more important.