postlink-68k-appl:  Axe superfluous 'CODE' 1 name
This tag provides zero information *at best*, since 'CODE' 1 is *always*
the main code resource.  An exception is the rare instance of deliberate
obfuscation intended to frustrate reverse-engineering attempts -- but in
this case, the obfuscating party is clearly *not* going to be helpful to
the effort by pointing to the real main entry point, but could very well
indicate the *wrong* one!  So at worst, the name bears *negative* value.