amicus -> frontend-common:  Move tempfile_location()
Since this function is no longer owned by the amicus project, a new name
is needed to tag the temporary directories it creates.  Although it's an
easy patch to substitute one project name for another, "frontend-common"
has various defects:  It's too long, "common" doesn't convey anything of
value, and "frontend" is actually misleading, since the .works structure
is shared by frontend and backend alike.  But "displayfs" in fact works!